move into util
